+++ apps/builder/app/builder/features/sidebar-left/panels/pages/social-utils.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-irregular-whitespace */
 import { expect, test } from "@jest/globals";
-import { truncateMetaText, truncateUrl } from "./search-preview";
+import { truncateMetaText, truncateUrl } from "./social-utils";
 
 test("Truncates description", () => {
   expect(
